+++ src/virt.rs
@@ -213,23 +213,22 @@ pub mod dispatch {
 use std::path::PathBuf;
 use trussed::{
     types::Bytes,
-    virt::{self, Client, Filesystem, Ram, StoreProvider},
+    virt::{self, Client, StorageConfig, StoreConfig},
 };
 
 /// Client type using a dispatcher with the backends required by opcard
-pub type VirtClient<S> = Client<S, dispatch::Dispatch>;
+pub type VirtClient<'a> = Client<'a, dispatch::Dispatch>;
 
 /// Run a client using a provided store
-pub fn with_client<S, R, F>(store: S, client_id: &str, f: F) -> R
+pub fn with_client<R, F>(store: StoreConfig, client_id: &str, f: F) -> R
 where
-    F: FnOnce(VirtClient<S>) -> R,
-    S: StoreProvider,
+    F: FnOnce(VirtClient<'_>) -> R,
 {
     #[allow(clippy::unwrap_used)]
     virt::with_platform(store, |platform| {
         platform.run_client_with_backends(
             client_id,
-            dispatch::Dispatch::with_hw_key(Bytes::from_slice(b"some bytes").unwrap()),
+            dispatch::Dispatch::with_hw_key(Bytes::from(b"some bytes")),
             dispatch::BACKENDS,
             f,
         )
@@ -240,17 +239,22 @@ where
 /// using storage backed by a file
 pub fn with_fs_client<P, R, F>(internal: P, client_id: &str, f: F) -> R
 where
-    F: FnOnce(VirtClient<Filesystem>) -> R,
+    F: FnOnce(VirtClient<'_>) -> R,
     P: Into<PathBuf>,
 {
-    with_client(Filesystem::new(internal), client_id, f)
+    let store = StoreConfig {
+        internal: StorageConfig::filesystem(internal.into()),
+        external: StorageConfig::ram(),
+        volatile: StorageConfig::ram(),
+    };
+    with_client(store, client_id, f)
 }
 
 /// Run the backend with the extensions required by opcard
 /// using a RAM file storage
 pub fn with_ram_client<R, F>(client_id: &str, f: F) -> R
 where
-    F: FnOnce(VirtClient<Ram>) -> R,
+    F: FnOnce(VirtClient<'_>) -> R,
 {
-    with_client(Ram::default(), client_id, f)
+    with_client(StoreConfig::ram(), client_id, f)
 }